Add navigation layout changes for onboarding additional wallets
